fix: 매니저 탈퇴 동시성 문제 해결

+++ src/main/java/site/codemonster/comon/domain/auth/service/MemberService.java
@@ -78,9 +78,14 @@ public class MemberService {
         teamRecruitLowService.deleteByMemberId(memberId);
         teamApplyLowService.deleteTeamAppliesByMemberId(memberId);
 
-        List<Team> teamsManagedByMember = teamLowService.findByTeamMangerIdForDelete(memberId);
+        List<Team> teamsManagedByMember = teamLowService.findByTeamMangerId(memberId);
+
         for (Team team : teamsManagedByMember) {
             Long teamId = team.getTeamId();
+
+            List<TeamMember> teamManagers = teamMemberLowService.findTeamManagerByTeamIdForUpdate(teamId);
+
+            if (teamManagers.size() > 1) continue;
 
             articleLowService.deleteByTeamId(teamId);
